feat : vue reset

+++ app/src/main/java/com/example/partagedefrais/dao/DepenseDao.java
@@ -255,9 +255,9 @@ public class DepenseDao {
         if (c.getCount() == 0) {
             aRenvoyer = null;
         } else {
-//            c.moveToFirst();
+            // c.moveToFirst();
             // on initialise l'instance Depense avec les valeurs des colonnes
-            aRenvoyer = new Depense(c.getInt(COLONNE_CLE));
+            aRenvoyer = new Depense(c.getLong(COLONNE_CLE));
             aRenvoyer.setNom(c.getString(COLONNE_NOM));
             aRenvoyer.setMontant(c.getDouble(COLONNE_MONTANT));
         }
